RFC: WIP: Honor user namespace privileges with zfs unallow
Regular users can only remove permissions from themselves in
addition to requiring the allow permission to do so. It
makes more sense for privileged users in a user namespace
to be able to manage permissions of all users of that
namespace.
Thus, when the user has CAP_SYS_ADMIN in their current
namespace, use the same check as for 'zfs allow' instead.

TODO: This should add tests / extend the ones from the
      previous commit.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>